+++ RELEASES.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# 0.11.0
 ## Changed
 - Bumped MSRV to `1.85.0`
 - Bumped dependencies to `bls12_381 0.9`, `ff 0.14`, `group 0.14`, `rand_core 0.10`.